cizzo_practice/requirements.txt
Babel==2.3.4
blinker==1.4
coverage==4.0.3
decorator==4.0.9
defusedxml==0.4.1
Flask==0.10.1
Flask-Babel==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.9.1
python3-openid==3.0.10
pytz==2016.4
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
Werkzeug==0.11.9
Whoosh==2.7.4
WTForms==2.1
